refactor(sequence-conversion): move mod-dictionary projection into serializers
