+++ mzLib/Omics/SequenceConversion/SequenceConversionExtensions.cs
@@ -193,7 +193,8 @@ public static class SequenceConversionExtensions
             return new Dictionary<int, Modification>();
         }
 
-        return canonical.Value.ToOneIsNterminusModificationDictionary(knownMods, mode, warnings);
+        var serializer = ResolveSerializerForCanonical(service, canonical.Value);
+        return serializer.ToOneIsNterminusModificationDictionary(canonical.Value, knownMods, warnings, mode);
     }
 
     public static Dictionary<int, Modification> ToOneIsNterminusModificationDictionary(
@@ -203,45 +204,23 @@ public static class SequenceConversionExtensions
         ConversionWarnings? warnings = null)
     {
         ArgumentNullException.ThrowIfNull(knownMods);
+        var serializer = ResolveSerializerForCanonical(SequenceConversionService.Default, canonical);
+        return serializer.ToOneIsNterminusModificationDictionary(canonical, knownMods, warnings, mode);
+    }
 
-        var allModsOneIsNterminus = new Dictionary<int, Modification>();
-        foreach (var canonicalModification in canonical.Modifications)
+    private static ISequenceSerializer ResolveSerializerForCanonical(SequenceConversionService service, CanonicalSequence canonical)
+    {
+        var sourceFormat = canonical.SourceFormat;
+        if (!string.IsNullOrWhiteSpace(sourceFormat))
         {
-            var index = ResolveOneIsNterminusIndex(canonicalModification, canonical.BaseSequence.Length);
-            var targetResidue = ResolveTargetResidue(canonicalModification, canonical.BaseSequence);
-
-            if (!TryResolveKnownModification(canonicalModification, targetResidue, knownMods, out var resolved))
+            var serializer = service.GetSerializer(sourceFormat);
+            if (serializer != null)
             {
-                if (mode == SequenceConversionHandlingMode.ThrowException)
-                {
-                    throw new SequenceConversionException(
-                        $"Unable to resolve modification {canonicalModification}",
-                        ConversionFailureReason.IncompatibleModifications,
-                        new[] { canonicalModification.ToString() });
-                }
-
-                warnings?.AddWarning($"Unable to resolve modification {canonicalModification}; skipping.");
-                continue;
+                return serializer;
             }
-
-            if (allModsOneIsNterminus.ContainsKey(index))
-            {
-                if (mode == SequenceConversionHandlingMode.ThrowException)
-                {
-                    throw new SequenceConversionException(
-                        $"Multiple modifications map to OneIsNterminus index {index}.",
-                        ConversionFailureReason.InvalidSequence,
-                        new[] { canonicalModification.ToString() });
-                }
-
-                warnings?.AddWarning($"Multiple modifications map to OneIsNterminus index {index}; keeping the first entry.");
-                continue;
-            }
-
-            allModsOneIsNterminus.Add(index, resolved!);
         }
 
-        return allModsOneIsNterminus;
+        return MzLibSequenceSerializer.Instance;
     }
 
     private static IDictionary<int, List<Modification>> SelectModDictionary(
@@ -280,244 +259,6 @@ public static class SequenceConversionExtensions
         }
 
         return modifications[0];
-    }
-
-    private static int ResolveOneIsNterminusIndex(CanonicalModification canonicalModification, int baseSequenceLength)
-    {
-        switch (canonicalModification.PositionType)
-        {
-            case ModificationPositionType.NTerminus:
-                return 1;
-
-            case ModificationPositionType.CTerminus:
-                return baseSequenceLength + 2;
-
-            case ModificationPositionType.Residue:
-                if (!canonicalModification.ResidueIndex.HasValue ||
-                    canonicalModification.ResidueIndex.Value < 0 ||
-                    canonicalModification.ResidueIndex.Value >= baseSequenceLength)
-                {
-                    throw new SequenceConversionException(
-                        $"Residue index {canonicalModification.ResidueIndex?.ToString() ?? "null"} is invalid for base sequence length {baseSequenceLength}.",
-                        ConversionFailureReason.InvalidSequence,
-                        new[] { canonicalModification.ToString() });
-                }
-
-                return canonicalModification.ResidueIndex.Value + 2;
-
-            default:
-                throw new SequenceConversionException(
-                    $"Unsupported modification position type: {canonicalModification.PositionType}",
-                    ConversionFailureReason.InvalidSequence,
-                    new[] { canonicalModification.ToString() });
-        }
-    }
-
-    private static char? ResolveTargetResidue(CanonicalModification canonicalModification, string baseSequence)
-    {
-        if (canonicalModification.TargetResidue.HasValue)
-        {
-            return canonicalModification.TargetResidue.Value;
-        }
-
-        if (canonicalModification.PositionType == ModificationPositionType.NTerminus)
-        {
-            return baseSequence.Length > 0 ? baseSequence[0] : (char?)null;
-        }
-
-        if (canonicalModification.PositionType == ModificationPositionType.CTerminus)
-        {
-            return baseSequence.Length > 0 ? baseSequence[^1] : (char?)null;
-        }
-
-        if (canonicalModification.ResidueIndex.HasValue &&
-            canonicalModification.ResidueIndex.Value >= 0 &&
-            canonicalModification.ResidueIndex.Value < baseSequence.Length)
-        {
-            return baseSequence[canonicalModification.ResidueIndex.Value];
-        }
-
-        return null;
-    }
-
-    private static bool TryResolveKnownModification(
-        CanonicalModification canonicalModification,
-        char? targetResidue,
-        Dictionary<string, Modification> knownMods,
-        out Modification? resolved)
-    {
-        resolved = null;
-
-        if (canonicalModification.MzLibModification != null)
-        {
-            if (!string.IsNullOrWhiteSpace(canonicalModification.MzLibModification.IdWithMotif) &&
-                knownMods.TryGetValue(canonicalModification.MzLibModification.IdWithMotif, out var byIdWithMotif))
-            {
-                resolved = byIdWithMotif;
-                return true;
-            }
-
-            if (knownMods.Values.Contains(canonicalModification.MzLibModification))
-            {
-                resolved = canonicalModification.MzLibModification;
-                return true;
-            }
-        }
-
-        if (TryResolveByToken(canonicalModification.MzLibId, targetResidue, knownMods, out resolved))
-        {
-            return true;
-        }
-
-        if (TryResolveByToken(canonicalModification.OriginalRepresentation, targetResidue, knownMods, out resolved))
-        {
-            return true;
-        }
-
-        if (canonicalModification.UnimodId.HasValue &&
-            TryResolveByUnimodId(canonicalModification.UnimodId.Value, targetResidue, knownMods, out resolved))
-        {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static bool TryResolveByToken(
-        string? modToken,
-        char? targetResidue,
-        Dictionary<string, Modification> knownMods,
-        out Modification? resolved)
-    {
-        resolved = null;
-        if (string.IsNullOrWhiteSpace(modToken))
-        {
-            return false;
-        }
-
-        var trimmed = modToken.Trim();
-        if (knownMods.TryGetValue(trimmed, out var exactMatch))
-        {
-            resolved = exactMatch;
-            return true;
-        }
-
-        var splitIndex = trimmed.IndexOf(':');
-        if (splitIndex > 0 && splitIndex < trimmed.Length - 1)
-        {
-            var tokenWithoutPrefix = trimmed.Substring(splitIndex + 1).Trim();
-            if (knownMods.TryGetValue(tokenWithoutPrefix, out var withoutPrefixMatch))
-            {
-                resolved = withoutPrefixMatch;
-                return true;
-            }
-        }
-
-        if (TryParseUnimodToken(trimmed, out var unimodId) &&
-            TryResolveByUnimodId(unimodId, targetResidue, knownMods, out resolved))
-        {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static bool TryResolveByUnimodId(
-        int unimodId,
-        char? targetResidue,
-        Dictionary<string, Modification> knownMods,
-        out Modification? resolved)
-    {
-        resolved = null;
-
-        var candidates = knownMods.Values
-            .Where(modification => modification != null && MatchesUnimodId(modification, unimodId))
-            .Distinct()
-            .ToList();
-
-        if (candidates.Count == 0)
-        {
-            return false;
-        }
-
-        if (candidates.Count == 1)
-        {
-            resolved = candidates[0];
-            return true;
-        }
-
-        if (!targetResidue.HasValue)
-        {
-            return false;
-        }
-
-        var residueMatches = candidates
-            .Where(candidate =>
-            {
-                var target = candidate.Target?.ToString();
-                return !string.IsNullOrWhiteSpace(target) &&
-                       target.Contains(targetResidue.Value.ToString(), StringComparison.OrdinalIgnoreCase);
-            })
-            .Distinct()
-            .ToList();
-
-        if (residueMatches.Count != 1)
-        {
-            return false;
-        }
-
-        resolved = residueMatches[0];
-        return true;
-    }
-
-    private static bool MatchesUnimodId(Modification modification, int unimodId)
-    {
-        if (modification.DatabaseReference != null)
-        {
-            foreach (var databaseReference in modification.DatabaseReference)
-            {
-                if (!databaseReference.Key.Equals("UNIMOD", StringComparison.OrdinalIgnoreCase))
-                {
-                    continue;
-                }
-
-                foreach (var token in databaseReference.Value)
-                {
-                    if (TryParseUnimodToken(token, out var parsedId) && parsedId == unimodId)
-                    {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        return TryParseUnimodToken(modification.Accession, out var accessionId) && accessionId == unimodId;
-    }
-
-    private static bool TryParseUnimodToken(string? token, out int unimodId)
-    {
-        unimodId = 0;
-        if (string.IsNullOrWhiteSpace(token))
-        {
-            return false;
-        }
-
-        var trimmed = token.Trim();
-        var lastColon = trimmed.LastIndexOf(':');
-        if (lastColon >= 0 && lastColon < trimmed.Length - 1)
-        {
-            var prefix = trimmed.Substring(0, lastColon).Trim();
-            if (!string.IsNullOrWhiteSpace(prefix) &&
-                !prefix.Equals("UNIMOD", StringComparison.OrdinalIgnoreCase) &&
-                !prefix.EndsWith("UNIMOD", StringComparison.OrdinalIgnoreCase))
-            {
-                return false;
-            }
-
-            trimmed = trimmed.Substring(lastColon + 1).Trim();
-        }
-
-        return int.TryParse(trimmed, out unimodId);
     }
 
     private static bool IsNTerminal(string? locationRestriction, int index)
